More code cut-paste.
